requirements.txt
pydantic-ai
openai
google-gepydantic-ai
python-telegram-bot
icalendar
dateparser
python-dateutil
